Integrated the Sesame TCK with the embedded federation for use in CI validation of the scale-out quads mode query.
This commit also fixes a few places in the query plan generation, including:

- SliceOp wraps the query plan so the top-level operator always runs on the query controller.
- Access path flags set the READONLY bit.
- Cost estimates needed to use the glocal index view (remoteAccessPath=true).

As of now, all but 4 of the tests in the SPARQL test suite are passing against the embedded federation in quads mode.  The four tests which are still failing are:

- "Union is not optional".
- "graph-07"
- "graph-11"
- "SELECT DISTINCT *"

git-svn-id: svn+ssh://svn.code.sf.net/p/bigdata/code/branches/QUADS_QUERY_BRANCH@3767 8f7bc0f5-282e-0410-95e3-8d296e9bb460

+++ bigdata/src/test/com/bigdata/service/EmbeddedFederation.java
@@ -932,13 +932,17 @@ public class EmbeddedFederation<T> extends AbstractScaleOutFederation<T> {
 
         if (metadataService != null) {
 
-            // the file flagging this as the MDS rather than a DS.
-            final File tmp = new File(metadataService.getResourceManager()
-                    .getDataDir(), EmbeddedFederation.MDS);
+            if (!isTransient()) {
 
-            if(!tmp.delete()) {
+                // the file flagging this as the MDS rather than a DS.
+                final File tmp = new File(metadataService.getResourceManager()
+                        .getDataDir(), EmbeddedFederation.MDS);
 
-                log.warn(ERR_COULD_NOT_DELETE + tmp);
+                if (!tmp.delete()) {
+
+                    log.warn(ERR_COULD_NOT_DELETE + tmp);
+
+                }
 
             }
